Small changes to make tests work compactly

+++ reportlab/test/runAll.py
@@ -2,24 +2,24 @@
 #copyright ReportLab Inc. 2000-2001
 #see license.txt for license details
 #history http://cvs.sourceforge.net/cgi-bin/cvsweb.cgi/reportlab/test/runAll.py?cvsroot=reportlab
-#$Header: /tmp/reportlab/reportlab/test/runAll.py,v 1.13 2004/03/22 13:21:06 rgbecker Exp $
+#$Header: /tmp/reportlab/reportlab/test/runAll.py,v 1.14 2004/03/22 18:08:50 rgbecker Exp $
 """Runs all test files in all subfolders.
 """
 import os, glob, sys, string, traceback
 from reportlab.test import unittest
 from reportlab.test.utils import GlobDirectoryWalker
 
-def makeSuite(folder, exclude=[],nonImportable=[]):
+def makeSuite(folder, exclude=[],nonImportable=[],pattern='test_*.py'):
     "Build a test suite of all available test files."
 
     allTests = unittest.TestSuite()
 
     if os.path.isdir(folder): sys.path.insert(0, folder)
-    for filename in GlobDirectoryWalker(folder, 'test_*.py'):
+    for filename in GlobDirectoryWalker(folder, pattern):
         modname = os.path.splitext(os.path.basename(filename))[0]
         if modname not in exclude:
             try:
-                module = __import__(modname)
+                exec 'import %s as module' % modname
                 allTests.addTest(module.makeSuite())
             except:
                 tt, tv, tb = sys.exc_info()[:]
@@ -35,6 +35,8 @@ def main():
         folder = os.path.dirname(__file__)
     except:
         folder = os.path.dirname(sys.argv[0]) or os.getcwd()
+    from reportlab.lib.utils import isSourceDistro
+    haveSRC = isSourceDistro()
 
     # special case for reportlab/test directory - clean up
     # all PDF & log files before starting run.  You don't
@@ -46,10 +48,14 @@ def main():
                 os.remove(filename)
 
     NI = []
-    testSuite = makeSuite(folder,nonImportable=NI)
+    testSuite = makeSuite(folder,nonImportable=NI,pattern='test_*.py'+(not haveSRC and 'c' or ''))
     unittest.TextTestRunner().run(testSuite)
-    for filename in GlobDirectoryWalker(folder, '*.pyc'):
-        os.remove(filename)
+    if haveSRC:
+        for filename in GlobDirectoryWalker(folder, '*.pyc'):
+            try:
+                os.remove(filename)
+            except:
+                pass
     if NI:
         sys.stderr.write('\n###################### the following tests could not be imported\n')
         for f,tb in NI:
